requirements.txt
AWSIoTPythonSDK==1.4.9
RPi.GPIO==0.7.0
